Added security and permissions to teachers

+++ app/controllers/api/v1/teachers_controller.rb
@@ -1,12 +1,12 @@
 class Api::V1::TeachersController < ApplicationController
-  # before_action :doorkeeper_authorize!
-  before_action :set_teacher, only: %i[show update destroy availability]
+  before_action :doorkeeper_authorize!
+  before_action :set_teacher, only: %i[show update destroy availability appointments]
+  before_action :set_teachers, only: %i[index]
+  before_action :set_current_user, only: %i[appointments create update destroy]
 
   # GET /teachers
   def index
-    @teachers = Teacher.all
-
-    render json: @teachers
+    render :index
   end
 
   # GET /teachers/1
@@ -14,41 +14,70 @@ class Api::V1::TeachersController < ApplicationController
     render :show
   end
 
-  # GET /teacher/1/availability/2020-01-01
+  # GET /teachers/1/availability/2020-01-01
   def availability
-    @appointments = Appointment.where('teacher_id = ? AND scheduled_for BETWEEN ? AND ?', @teacher.id, params[:date].to_date, params[:date].to_date.next_day(1))
+    if params[:date].to_date > Time.now.utc.to_date
+      @appointments = Appointment.where('teacher_id = ? AND scheduled_for BETWEEN ? AND ?', @teacher.id, params[:date].to_date, params[:date].to_date.next_day(1))
 
-    @availability = { 'availability' => (8..19).to_a - @appointments.map { |item| item.scheduled_for.hour }.to_a }
+      @availability = { 'availability' => (8..19).to_a - @appointments.map { |item| item.scheduled_for.hour }.to_a }
+    else
+      @availability = { 'availability' => [] }
+    end
 
     render json: @availability
   end
 
+  # GET /teachers/1/appointments
+  def appointments
+    if @current_user.admin
+      @appointments = Appointment
+        .select(:id, :scheduled_for, 'users.fullname as user_fullname', 'users.email as user_email', 'CASE WHEN scheduled_for > timezone(\'utc\', now()) THEN 1 ELSE 0 END as status')
+        .joins(:user).where(teacher_id: @teacher.id)
+        .order(scheduled_for: :asc)
+
+      render json: @appointments
+    else
+      render json: return_error_message(403), status: :forbidden
+    end
+  end
+
   # POST /teachers
   def create
-    @teacher = Teacher.new(teacher_params)
-
-    if @teacher.save
-      render json: @teacher, status: :created
+    if @current_user.admin
+      @teacher = Teacher.new(teacher_params)
+      if @teacher.save
+        render json: @teacher, status: :created
+      else
+        render json: @teacher.errors, status: :unprocessable_entity
+      end
     else
-      render json: @teacher.errors, status: :unprocessable_entity
+      render json: return_error_message(403), status: :forbidden
     end
   end
 
   # PATCH/PUT /teachers/1
   def update
-    if @teacher.update(teacher_params)
-      render json: @teacher
+    if @current_user.admin
+      if @teacher.update(teacher_params)
+        render :show
+      else
+        render json: @teacher.errors, status: :unprocessable_entity
+      end
     else
-      render json: @teacher.errors, status: :unprocessable_entity
+      render json: return_error_message(403), status: :forbidden
     end
   end
 
   # DELETE /teachers/1
   def destroy
-    Appointment.where(teacher_id: @teacher.id).destroy_all
-    @teacher.destroy
-
-    render json: @teacher
+    if @current_user.admin
+      Appointment.where(teacher_id: @teacher.id).destroy_all
+      @teacher.destroy
+  
+      render :show
+    else
+      render json: return_error_message(403), status: :forbidden
+    end
   end
 
   private
@@ -56,6 +85,14 @@ class Api::V1::TeachersController < ApplicationController
   # Use callbacks to share common setup or constraints between actions.
   def set_teacher
     @teacher = Teacher.find(params[:id])
+  end
+
+  def set_teachers
+    @teachers = Teacher.all
+  end
+
+  def set_current_user
+    @current_user = current_user
   end
 
   # Only allow a trusted parameter "white list" through.
